Limit unneeded emails, include ticket in email
Don't send email to customer when they reply to their own ticket. When admin replies to a ticket, if this admin's user email is the same as the wpsc Support Ticket setting's send-to email, then do not send email notification to this admin for this reply which he wrote. Also, include ticket message body in the emails, with a note at bottom to: View the entire ticket and give a reply, with a link back to the main support ticket page.